refactor: group built-in forge backends
Move the built-in forge implementations under lua/forge/backends so
forge-specific code is separated from the core workflow modules while
keeping compatibility wrappers for existing require paths.
