+++ lua/forge/init.lua
@@ -173,7 +173,10 @@ local function resolve_source(name)
   if sources[name] then
     return sources[name]
   end
-  local ok, mod = pcall(require, 'forge.' .. name)
+  local ok, mod = pcall(require, 'forge.backends.' .. name)
+  if not ok then
+    ok, mod = pcall(require, 'forge.' .. name)
+  end
   if ok then
     sources[name] = mod
     return mod
